Use new lzo project

+++ src/utils/xrCompress/xrCompress.h
@@ -33,7 +33,7 @@ class xrCompressor
 	
 	void	GatherFiles			(LPCSTR folder);
 	
-	void	write_file_header	(LPCSTR file_name, const u32 &crc, const u32 &ptr, const u32 &size_real, const u32 &size_compressed);
+	void	write_file_header	(LPCSTR file_name, const u32 &crc, const u32 &ptr, const u32 &size_real, const size_t &size_compressed);
 	void	ClosePack			();
 	void	OpenPack			(LPCSTR tgt_folder, int num);
